+++ resources/views/listadocpref.blade.php
@@ -157,8 +157,11 @@
 
         $('#table_id tbody').on( 'click', 'button#delete', function () {
             var data = table.row( $(this).parents('tr') ).data();
-            DeleteThis(data['id']);
-            table.row($(this).parents('tr')).remove().draw();
+            var confirmmssg = confirm("Sicuro di voler eliminare il documento?");
+            if (confirmmssg ) {
+                DeleteThis(data['id']);
+                table.row($(this).parents('tr')).remove().draw();
+            }
         } );
 
         $('#table_id tbody').on( 'click', 'button#delTag', function () {
@@ -189,10 +192,9 @@
 
     function DeleteThis( id )
     {
-        var confirmmssg = confirm("Sicuro di voler eliminare il documento?");
         var url = '{{ route('delDoc', ':id') }}';
         url = url.replace(':id', id);
-        if (confirmmssg ){
+
             $.ajaxSetup({
                 headers: {
                     'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')
@@ -202,7 +204,7 @@
                 type: "delete",
                 url: url,
             });
-        }
+
     }
 
     function DeleteTag( id )
